Re-Refactored filesys
