+++ bigtoolbox/io/src/main/java/com/incplusplus/bigtoolbox/io/filesys/Entry.java
@@ -1,4 +1,4 @@
-package com.incplusplus.bigtoolbox.io;
+package com.incplusplus.bigtoolbox.io.filesys;
 
 public abstract class Entry
 {
