+++ bigtoolbox/io/src/main/java/com/incplusplus/bigtoolbox/io/filesys/Record.java
@@ -1,4 +1,4 @@
-package com.incplusplus.bigtoolbox.io;
+package com.incplusplus.bigtoolbox.io.filesys;
 
 import java.io.File;
 import java.io.FileNotFoundException;
