--- bigtoolbox/io/src/test/java/com/incplusplus/bigtoolbox/io/DirectoryTest.java
@@ -1,7 +1,0 @@
-package com.incplusplus.bigtoolbox.io;
-
-//TODO Complete this
-class DirectoryTest
-{
-
-}
