+++ bigtoolbox/io/src/test/java/com/incplusplus/bigtoolbox/io/filesys/FileRelatedToolsTest.java
@@ -1,4 +1,4 @@
-package com.incplusplus.bigtoolbox.io;
+package com.incplusplus.bigtoolbox.io.filesys;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
